# perf: use tiny cache for best chainlock in CbTx instead running heavy ReadBlockFromDisk